requirements.txt
numpy==1.22.4
opencv_python==4.7.0.72
Pillow==9.5.0
scikit_image==0.19.3